Major refactor to TelegramClient to prepare for commands requiring verification.
