+++ src/telegram-client/telegram-client-mock.ts
@@ -1,21 +1,11 @@
-import { BotCommand } from "../bot-commands/bot-command";
 import { ITelegramClient } from "./i-telegram-client";
 import { ITelegramClientListener } from "./i-telegram-client-listener";
 
 export class TelegramClientMock implements ITelegramClient {
 
-  public readonly commands = new Map<string, BotCommand>();
   public readonly botname = "testbot";
 
-  public async executeCommand(msg: any, match: string[], botCommand: BotCommand): Promise<string> {
-    return "";
-  }
-
   public setOnAnyText(action: (msg: any, match: string[]) => string[]): void {
-    // Don't do anything, this is a mock.
-  }
-
-  public async registerCommand(command: BotCommand): Promise<void> {
     // Don't do anything, this is a mock.
   }
 
@@ -29,5 +19,21 @@ export class TelegramClientMock implements ITelegramClient {
 
   public async deleteMessage(chatId: number, messageId: number): Promise<any> {
     // Don't do anything, this is a mock.
+  }
+
+  public getBotUsername(): Promise<string> {
+    return Promise.resolve("test");
+  }
+
+  public setOnRegex(regExp: RegExp, action: (msg: any, match: string[]) => void): void {
+    // Don't do anything, this is a mock.
+  }
+
+  public getChatAdministrators(chatId: number): Promise<any[]> {
+    return Promise.resolve([{
+      user: {
+        id: 0,
+      },
+    }]);
   }
 }
